Change license: GNU GPL v3.0 -> Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 Internationall
License Notice

This documentation, document, and the ideas contained within it are licensed under the Creative Commons Attribution-NoDerivatives-NonCommercial 4.0 International License. This means you are free to share (copy and redistribute the material in any medium or format) as long as you provide appropriate credit. However, you may not modify, adapt, or create derivative works based on this material, nor use it for commercial purposes. 

For full license details, visit: https://creativecommons.org/licenses/by-nc-nd/4.0/. 

The age-based versioning system itself is free to use and implement by anyone. However, this document and its explanation are protected under the aforementioned license.

Signed-off-by: Farhan Kurnia Pratama <farhnkrnapratma@gmail.com>